Second commit

+++ src/BmiService.java
@@ -2,7 +2,7 @@ public class BmiService {
 
     public double calculate(double weight, double height) {
         double result;
-        result = (double) (weight / (height*height));
+        result = weight / (height * height);
         return result;
     }
 }
